swap functions moved to ArrayUtil
